# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Espaço Girafinha reservation management API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: reservations
    description: Reservation management
  - name: dashboard
    description: Dashboard and analytics

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /reservations:
    get:
      operationId: listReservations
      tags: [reservations]
      summary: List all reservations
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [paid, partial, unpaid]
        - name: serviceType
          in: query
          schema:
            $ref: "#/components/schemas/ServiceType"
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of reservations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Reservation"

    post:
      operationId: createReservation
      tags: [reservations]
      summary: Create a new reservation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReservationBody"
      responses:
        "201":
          description: Reservation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /reservations/{id}:
    get:
      operationId: getReservation
      tags: [reservations]
      summary: Get a reservation by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reservation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateReservation
      tags: [reservations]
      summary: Update a reservation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateReservationBody"
      responses:
        "200":
          description: Updated reservation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deleteReservation
      tags: [reservations]
      summary: Delete a reservation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [dashboard]
      summary: Get dashboard statistics
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"

  /dashboard/upcoming:
    get:
      operationId: getUpcomingReservations
      tags: [dashboard]
      summary: Get upcoming reservations
      responses:
        "200":
          description: Upcoming reservations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Reservation"

  /reservations/{id}/tasks:
    get:
      operationId: listTasks
      tags: [reservations]
      summary: List tasks for a reservation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"

    post:
      operationId: createTask
      tags: [reservations]
      summary: Create a new task for a reservation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskBody"
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"

  /tasks/{taskId}:
    patch:
      operationId: updateTask
      tags: [reservations]
      summary: Update a task
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskBody"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"

    delete:
      operationId: deleteTask
      tags: [reservations]
      summary: Delete a task
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /tasks/summary:
    get:
      operationId: getTasksSummary
      tags: [reservations]
      summary: Get a summary of task progress for many reservations
      parameters:
        - name: reservationIds
          in: query
          required: true
          description: Comma-separated list of reservation IDs
          schema:
            type: string
      responses:
        "200":
          description: Task summaries indexed by reservation ID
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TaskSummary"

  /reports:
    get:
      operationId: getReportsData
      tags: [dashboard]
      summary: Get reports and analytics data
      parameters:
        - name: year
          in: query
          schema:
            type: integer
        - name: month
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Reports data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportsData"

  /dashboard/calendar:
    get:
      operationId: getCalendarReservations
      tags: [dashboard]
      summary: Get reservations grouped by date for calendar view
      parameters:
        - name: year
          in: query
          required: true
          schema:
            type: integer
        - name: month
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Calendar data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CalendarDay"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    ServiceType:
      type: string
      enum: ["Festas no espaço", "Serviços externos", "Workshops"]

    Reservation:
      type: object
      properties:
        id:
          type: integer
        customerName:
          type: string
        phone:
          type: string
        eventDate:
          type: string
        eventTime:
          type: string
        pack:
          type: string
          enum: ["Aluguer do Espaço", "Pack Simples", "Pack com Decoração", "Pack Completo", "Decoração Externa", "Catering / Brunch", "Animação", "Aluguer de Insuflável", "Workshop Balões Nível 1", "Workshop Balões + Kit Inicial", "Só Espaço", "Espaço + Lanche", "Espaço + Decoração"]
        serviceType:
          $ref: "#/components/schemas/ServiceType"
        numChildren:
          type: integer
        childrenAges:
          type: string
        extras:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        totalPrice:
          type: number
        amountPaid:
          type: number
        remainingBalance:
          type: number
        paymentStatus:
          type: string
          enum: [paid, partial, unpaid]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - customerName
        - phone
        - eventDate
        - eventTime
        - pack
        - serviceType
        - numChildren
        - childrenAges
        - totalPrice
        - amountPaid
        - remainingBalance
        - paymentStatus
        - createdAt
        - updatedAt

    CreateReservationBody:
      type: object
      properties:
        customerName:
          type: string
        phone:
          type: string
        eventDate:
          type: string
        eventTime:
          type: string
        pack:
          type: string
          enum: ["Aluguer do Espaço", "Pack Simples", "Pack com Decoração", "Pack Completo", "Decoração Externa", "Catering / Brunch", "Animação", "Aluguer de Insuflável", "Workshop Balões Nível 1", "Workshop Balões + Kit Inicial", "Só Espaço", "Espaço + Lanche", "Espaço + Decoração"]
        serviceType:
          $ref: "#/components/schemas/ServiceType"
        numChildren:
          type: integer
        childrenAges:
          type: string
        extras:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        totalPrice:
          type: number
        amountPaid:
          type: number
      required:
        - customerName
        - phone
        - eventDate
        - eventTime
        - pack
        - numChildren
        - childrenAges
        - totalPrice
        - amountPaid

    UpdateReservationBody:
      type: object
      properties:
        customerName:
          type: string
        phone:
          type: string
        eventDate:
          type: string
        eventTime:
          type: string
        pack:
          type: string
          enum: ["Aluguer do Espaço", "Pack Simples", "Pack com Decoração", "Pack Completo", "Decoração Externa", "Catering / Brunch", "Animação", "Aluguer de Insuflável", "Workshop Balões Nível 1", "Workshop Balões + Kit Inicial", "Só Espaço", "Espaço + Lanche", "Espaço + Decoração"]
        serviceType:
          $ref: "#/components/schemas/ServiceType"
        numChildren:
          type: integer
        childrenAges:
          type: string
        extras:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        totalPrice:
          type: number
        amountPaid:
          type: number

    DashboardStats:
      type: object
      properties:
        totalReservations:
          type: integer
        totalRevenue:
          type: number
        totalPaid:
          type: number
        totalPending:
          type: number
        upcomingCount:
          type: integer
        paidCount:
          type: integer
        partialCount:
          type: integer
        unpaidCount:
          type: integer
      required:
        - totalReservations
        - totalRevenue
        - totalPaid
        - totalPending
        - upcomingCount
        - paidCount
        - partialCount
        - unpaidCount

    Task:
      type: object
      properties:
        id:
          type: integer
        reservationId:
          type: integer
        title:
          type: string
        completed:
          type: boolean
        sortOrder:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - reservationId
        - title
        - completed
        - sortOrder
        - createdAt

    CreateTaskBody:
      type: object
      properties:
        title:
          type: string
      required:
        - title

    UpdateTaskBody:
      type: object
      properties:
        title:
          type: string
        completed:
          type: boolean

    TaskSummary:
      type: object
      properties:
        reservationId:
          type: integer
        total:
          type: integer
        completed:
          type: integer
      required:
        - reservationId
        - total
        - completed

    MonthlyData:
      type: object
      properties:
        month:
          type: string
        year:
          type: integer
        monthNum:
          type: integer
        reservationCount:
          type: integer
        revenue:
          type: number
        paid:
          type: number
        pending:
          type: number
      required:
        - month
        - year
        - monthNum
        - reservationCount
        - revenue
        - paid
        - pending

    ServiceTypeStats:
      type: object
      properties:
        serviceType:
          $ref: "#/components/schemas/ServiceType"
        count:
          type: integer
        percentage:
          type: number
        revenue:
          type: number
      required:
        - serviceType
        - count
        - percentage
        - revenue

    PaymentStatusStats:
      type: object
      properties:
        status:
          type: string
          enum: [paid, partial, unpaid]
        count:
          type: integer
        percentage:
          type: number
        revenue:
          type: number
        pending:
          type: number
      required:
        - status
        - count
        - percentage
        - revenue
        - pending

    OccupancyStats:
      type: object
      properties:
        bookedDays:
          type: integer
        fullDays:
          type: integer
        usedSlots:
          type: integer
        availableSlots:
          type: integer
        maxSlots:
          type: integer
        occupancyRate:
          type: number
      required:
        - bookedDays
        - fullDays
        - usedSlots
        - availableSlots
        - maxSlots
        - occupancyRate

    ExtraStats:
      type: object
      properties:
        extra:
          type: string
        count:
          type: integer
        revenue:
          type: number
      required:
        - extra
        - count
        - revenue

    PackStats:
      type: object
      properties:
        pack:
          type: string
        count:
          type: integer
        percentage:
          type: number
        revenue:
          type: number
      required:
        - pack
        - count
        - percentage
        - revenue

    ReportsData:
      type: object
      properties:
        selectedMonth:
          type: object
          properties:
            reservationCount:
              type: integer
            revenue:
              type: number
            paid:
              type: number
            pending:
              type: number
            avgRevenuePerBooking:
              type: number
          required:
            - reservationCount
            - revenue
            - paid
            - pending
            - avgRevenuePerBooking
        previousMonth:
          type: object
          properties:
            reservationCount:
              type: integer
            revenue:
              type: number
          required:
            - reservationCount
            - revenue
        packStats:
          type: array
          items:
            $ref: "#/components/schemas/PackStats"
        occupancyStats:
          $ref: "#/components/schemas/OccupancyStats"
        serviceTypeStats:
          type: array
          items:
            $ref: "#/components/schemas/ServiceTypeStats"
        paymentStatusStats:
          type: array
          items:
            $ref: "#/components/schemas/PaymentStatusStats"
        extraStats:
          type: array
          items:
            $ref: "#/components/schemas/ExtraStats"
        monthlyTrend:
          type: array
          items:
            $ref: "#/components/schemas/MonthlyData"
        bestMonth:
          type: object
          properties:
            month:
              type: string
            revenue:
              type: number
          required:
            - month
            - revenue
        insights:
          type: array
          items:
            type: string
      required:
        - selectedMonth
        - previousMonth
        - occupancyStats
        - serviceTypeStats
        - paymentStatusStats
        - extraStats
        - packStats
        - monthlyTrend
        - bestMonth
        - insights

    CalendarDay:
      type: object
      properties:
        date:
          type: string
        reservations:
          type: array
          items:
            $ref: "#/components/schemas/Reservation"
      required:
        - date
        - reservations