# lib/api-zod/src/generated/types/serviceTypeStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Espaço Girafinha reservation management API
 * OpenAPI spec version: 0.1.0
 */
import type { ServiceType } from "./serviceType";

export interface ServiceTypeStats {
  serviceType: ServiceType;
  count: number;
  percentage: number;
  revenue: number;
}